review: int -> size_t.

+++ test/runtime_test.cc
@@ -121,7 +121,8 @@ TEST_P(TestVM, CloneUntilOutOfMemory) {
     // Prevent clone from droping out of scope and freeing memory.
     clones.push_back(std::move(clone));
   }
-  int min_clones = 1000;
+
+  size_t min_clones = 1000;
 #if defined(__has_feature)
 #if __has_feature(thread_sanitizer)
   min_clones = 100;
